+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-- Better pathfinding for warring triad unlocks
-- landing fixes for 3673
+- Experimental feature: qst will now use an aethernet shortcut as a hint for where to teleport to for the automatic teleporting thing. I've tested it a bit, but it probably still has issues, please let me know how that goes. -alydev
+- Feature: Notice and command to clear completion log -alydev